requirements.txt
pyswd>=0.0.2
